[Feature add] update vector decay gradient to fp32

+++ xopes/ops/lightning_attn/log_decay/log_decay_with_cumsum/dld_with_cumsum_triton.py
@@ -317,7 +317,9 @@ def compute_dld_with_cumsum_triton(
     if sum_option == -1:
         dld = torch.empty(b, n, h, device=dld_q.device, dtype=dld_q.dtype)
     else:
-        dld = torch.empty_like(dld_q)
+        dld = torch.empty_like(
+            dld_q, dtype=torch.float32
+        )  # for vector decay, we use fp32 dtype to compute
 
     # Determine if using final_state
     use_final_state = final_state is not None and dfinal_state is not None
